requirements.txt
pandas >=0.0.1
streamlit >=22.0.3
BeautifulSoup >=0.0.1
mysql-connector >= 8.0.36
requests >=0.0.1




 

 